Translated load_atom RTL instruction
To support atoms we also declare them as external constants,
and load them in the begging of the functions, as we do
for erlang constants.

Also we add them to dictionary to correlate them with the relocation values
extracted from the object files.

Finally atoms are added to relocation references needed by the loader,
as type 0.