Changed hipe bif which patches atoms.
In LLVM we use medium code model, in order to have 64-bit addresses for
constants as hipe does. This forces us to also use 64-bit addresses for atoms.
Hipe on the other hand uses 32-bit addresses for atoms.
So we changed the hipe bif that patches atoms (hipe_patch_insn).