Update elf64_format with .rodata section support
Now there is a new function get_label_list/1 to extract a list with all
constants and literals offsets in code from .rodata section (or an empty list
if no such section exists). We might use this to create proper mapping for an
rtl switch statement (we should build ConstMap and LabelMap by hand).