No load commands before the get_stacktrace call
We do not want to load the virtual regs (HP, P, NSP) in case of a
get_stacktrace call (assumed to be the first call in every flow starting from a
fail/unwind label). In that way we overcome the problem with no return struct
in case of a non-normal exit of a function (i.e. in case of an exception
throw).