Changed pipeline in hipe_main and fixed bug
Better integration of the LLVM back end in the HiPE pipeline.

Also, fixed a bug in hipe_llvm_liveness:find_dead_in_block/2. The list of the
variables that are being defined in a basic block must not have any duplicated
elements in order to guarantee the correctness of the liveness annotation.