Refactoring handling of closures with stack args
Remove hipe_bifs:llvm_expose_closure/0 which was used for exposing the
addresses of closures with arguments in the stack. Instead, we create a label
before any such call and save these labels to a special constant table
(table_closures) which is parsed from the object file with the
elf_format:get_symtab_list/1 function.